#910 - Whitespace can confuse StringMatchingRecommender
- Extracted key sanitizer interfaces into own files
- Moved trie code into a separate package
